requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==2.2.3
pandas==2.2.3
propcache==0.3.0
python-dateutil==2.9.0.post0
pytz==2025.1
RapidFuzz==3.12.2
requests==2.32.3
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
